Changed Rect to RectF in abstract Particle class
And also now use float values for the dx and dy values of the particle. This is to prevent the issue of casting to ints causing 0 values for dx and dy.

+++ app/src/main/java/jacobmahoney/guardianofearth/particles/Laser.java
@@ -8,7 +8,7 @@ public class Laser extends Particle {
 
     private Paint paint;
 
-    public Laser(int x, int y, int dx, int dy) {
+    public Laser(int x, int y, float dx, float dy) {
 
         super(x, y, dx, dy, 12, 12);
